refactor(relay): move short-replace non-delivery classifier out of giant session_relay_sink (#4046 S1r-1)
The P2 fix (90aeefe) added the pure classifier + its unit test inline in
session_relay_sink.rs, re-inflating that frozen #3016-hot giant to 1726 prod
LoC (> 1690 baseline). Relocate the classifier and its mutation-guard test into
a new sibling submodule `session_relay_sink/delivery_outcome_classify.rs`
(discord/mod.rs untouched — it is a giant too, so a new `mod` line there would
trip its own baseline). The match arm now calls
`delivery_outcome_classify::short_replace_non_delivery_error(&non_delivery)`.

Behavior/contract unchanged: FreshDelivered (confirmed cross-verb POST,
unreachable this stage) -> non-retriable Permanent, mirroring terminal_send's
conservative Skipped; every other non-delivery -> retriable Transient. The
"reverting the FreshDelivered arm to Transient fails the assert" mutation
property is preserved in the relocated test.

session_relay_sink.rs is back to 1681 prod LoC (<= 1690); the new submodule is
42 prod LoC (not a giant). audit_maintainability.py --check: rc=0. cargo/fmt and
generate_inventory_docs not run (orchestrator aggregate build owns those);
generated docs left for the aggregate step to regenerate.
