+++ src/services/discord/session_relay_sink.rs
@@ -31,6 +31,7 @@ use crate::services::provider::ProviderKind;
 use crate::services::session_backend::StreamLineState;
 use tracing::Instrument;
 
+mod delivery_outcome_classify;
 mod idle_jsonl;
 // #3960: orphaned `SessionBoundRelay` TUI-direct reclaim (producer-liveness TOCTOU).
 mod orphan_reclaim;
@@ -744,28 +745,16 @@ impl SessionBoundDiscordRelaySink {
                 );
                 Ok(SessionRelayDeliveryOutcome::Delivered)
             }
-            // #4046 S1r-1 P2: a `FreshDelivered` outcome is an UNREACHABLE cross-verb
-            // result here — `SendFresh` is not this stage's short-replace plan. It is a
-            // CONFIRMED POST (see `DeliveryOutcome::FreshDelivered` docs), so it must
-            // NEVER be surfaced as retriable `Transient`: a blind retry would duplicate
-            // the already-landed POST (dedup-less when `persistence_recorded == false`).
-            // The pure classifier maps it to the NON-retriable `Permanent` arm and
-            // claims no delivery — mirroring the control site
-            // `tmux_watcher/terminal_send.rs`, where `FreshDelivered { .. }` maps to the
-            // conservative non-retry `WatcherShortReplaceResult::Skipped`. Every other
-            // non-delivery here (ambiguous `PartialContinuationFailure` / transport Err,
-            // lost-acquire `Transient`, empty-body `Skipped`) is genuinely uncommitted
-            // (I2 — offset NOT advanced) → retriable `Transient`. Routed through the
-            // pure free fn so this retry classification is unit-testable; reverting the
-            // `FreshDelivered` arm there to `Transient` re-arms the duplicate-POST
-            // booby-trap this fix closed (see
-            // `fresh_delivered_short_replace_outcome_is_not_retriable`).
+            // #4046 S1r-1 P2: FreshDelivered is a confirmed cross-verb POST (unreachable
+            // here) → non-retriable; all other non-deliveries are uncommitted → retriable.
+            // Classified out-of-line so this frozen #3016 giant stays put (see
+            // `delivery_outcome_classify`, mirrors terminal_send's conservative Skipped).
             non_delivery @ (toc::DeliveryOutcome::FreshDelivered { .. }
             | toc::DeliveryOutcome::Transient { .. }
             | toc::DeliveryOutcome::Unknown { .. }
-            | toc::DeliveryOutcome::Skipped) => {
-                Err(short_replace_controller_non_delivery_error(&non_delivery))
-            }
+            | toc::DeliveryOutcome::Skipped) => Err(
+                delivery_outcome_classify::short_replace_non_delivery_error(&non_delivery),
+            ),
         }
     }
 
@@ -1173,40 +1162,6 @@ enum SessionRelayDeliveryOutcome {
     Delivered,
     SentButUncommitted,
     NotDelivered,
-}
-
-/// #4046 S1r-1 P2: classify a session-bound short-replace controller outcome that did
-/// NOT confirm a placeholder edit into a sink error, keeping the retry classification
-/// pure and unit-testable (the inline `Delivered`/`NotDelivered` arm in
-/// `deliver_short_replace_via_controller` carries metrics/tracing/observability side
-/// effects and cannot be exercised in isolation).
-///
-/// `FreshDelivered` is a CONFIRMED POST reached via an impossible cross-verb path
-/// (`SendFresh` is not this stage's short-replace plan). It MUST be non-retriable:
-/// surfacing it as `Transient` would let a blind retry duplicate the landed POST
-/// (dedup-less when `persistence_recorded == false`). This mirrors the control site
-/// `tmux_watcher/terminal_send.rs`, which maps `FreshDelivered` to the conservative
-/// non-retry `WatcherShortReplaceResult::Skipped`. Every other non-delivery here
-/// (ambiguous `PartialContinuationFailure` / transport Err, lost-acquire `Transient`,
-/// empty-body `Skipped`) is genuinely uncommitted (offset NOT advanced) → retriable
-/// `Transient`.
-fn short_replace_controller_non_delivery_error(outcome: &toc::DeliveryOutcome) -> RelaySinkError {
-    match outcome {
-        // Confirmed cross-verb POST → NON-retriable. NEVER `Transient` (a blind retry
-        // would duplicate the POST). Flipping this back to `Transient` re-arms the
-        // duplicate-POST booby-trap #4046 S1r-1 P2 closed.
-        toc::DeliveryOutcome::FreshDelivered { .. } => RelaySinkError::Permanent(
-            "session-bound short-replace controller returned cross-verb FreshDelivered \
-             (unreachable this stage); refusing retriable classification to prevent a \
-             duplicate POST"
-                .to_string(),
-        ),
-        // Ambiguous / failed / lost-acquire / empty-body: no confirmed POST, offset NOT
-        // advanced → retriable.
-        _ => RelaySinkError::Transient(
-            "session-bound short-replace controller delivery not confirmed".to_string(),
-        ),
-    }
 }
 
 #[async_trait]
@@ -3158,44 +3113,6 @@ mod tests {
             partial,
             toc::DeliveryOutcome::Unknown { fell_back: false }
         ));
-    }
-
-    /// #4046 S1r-1 P2 mutation guard: a confirmed cross-verb `FreshDelivered` outcome
-    /// must be classified NON-retriably (`RelaySinkError::Permanent`), never as
-    /// retriable `Transient` — a blind retry of a landed POST would duplicate the
-    /// message (dedup-less when `persistence_recorded == false`). This mirrors the
-    /// control site `tmux_watcher/terminal_send.rs` (`FreshDelivered` →
-    /// `WatcherShortReplaceResult::Skipped`, non-retry). Reverting the `FreshDelivered`
-    /// arm of `short_replace_controller_non_delivery_error` back to `Transient` FAILS
-    /// this assert.
-    #[test]
-    fn fresh_delivered_short_replace_outcome_is_not_retriable() {
-        // committed + persistence_recorded=false is the exact dedup-less duplicate-POST
-        // hazard the fix targets.
-        let fresh = toc::DeliveryOutcome::FreshDelivered {
-            committed_to: Some(7),
-            persistence_recorded: false,
-        };
-        let err = short_replace_controller_non_delivery_error(&fresh);
-        assert!(
-            matches!(err, RelaySinkError::Permanent(_)),
-            "FreshDelivered (confirmed POST) must be non-retriable Permanent, got {err:?}"
-        );
-
-        // Genuinely-uncommitted outcomes (offset NOT advanced) stay retriable Transient.
-        for uncommitted in [
-            toc::DeliveryOutcome::Transient {
-                retry_from_offset: 0,
-            },
-            toc::DeliveryOutcome::Unknown { fell_back: false },
-            toc::DeliveryOutcome::Skipped,
-        ] {
-            let err = short_replace_controller_non_delivery_error(&uncommitted);
-            assert!(
-                matches!(err, RelaySinkError::Transient(_)),
-                "a genuinely-uncommitted controller outcome must stay retriable Transient, got {err:?}"
-            );
-        }
     }
 
     // #3089 A2b (review-fix Medium-1): drive the flag-ON `deliver_response`
